adding search functionality using useEffect hook
