+++ src/components/Ingredients/Ingredients.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 
 import IngredientForm from './IngredientForm';
 import IngredientList from './IngredientList';
@@ -7,33 +7,36 @@ import Search from './Search';
 const Ingredients = () => {
   const [ userIngredients, setUserIngredients ] = useState([]);
 
-  // this hook gets called after and for every render cycle
-  // with the [] as a second parameter, useEffect is only run once
-  // very similar to 'componentDidMount' in class based components
-  useEffect(() => {
-    fetch('https://react-hooks-9a19a.firebaseio.com/ingredients.json')
-      .then( resp => {
-        return resp.json();
-      })
-      .then(data => {
-        const ingredientsArray = [];
-        for( const key in data ) {
-          ingredientsArray.push({
-            id: key,
-            title: data[key].title,
-            amount: data[key].amount
-          });
-        }
+  /** 
+   * this hook gets called after and for every render cycle
+   * with the [] as a second parameter, useEffect is only run once
+   * very similar to 'componentDidMount' in class based components
+   * 
+   * beause we are fetching ingredients in the search dmoponent, we don't need to fetch them here
+   */
+   // useEffect(() => {
+  //   fetch('https://react-hooks-9a19a.firebaseio.com/ingredients.json')
+  //     .then( resp => {
+  //       return resp.json();
+  //     })
+  //     .then(data => {
+  //       const ingredientsArray = [];
+  //       for( const key in data ) {
+  //         ingredientsArray.push({
+  //           id: key,
+  //           title: data[key].title,
+  //           amount: data[key].amount
+  //         });
+  //       }
 
-        setUserIngredients(ingredientsArray);
-      })
-  }, []);
+  //       setUserIngredients(ingredientsArray);
+  //     })
+  // }, []);
 
   // you can have as many 'useEffect' hooks as required
   // adding in a second parameter in the array will
   // make this useEffect run only if that parameter changes
   useEffect( ()=> {
-    console.log('rendering content: ', userIngredients);
   }, [userIngredients] )
 
   const addIngredientHandler = (ing) => {
@@ -48,9 +51,16 @@ const Ingredients = () => {
     });
   }
 
-  const removeIngredient = (ingId) => {
+  const removeIngredient = ( ingId ) => {
     setUserIngredients( prevState => prevState.filter( ing => ing.id !== ingId));
   }
+
+  /**
+   * remember - [] maked the hook only run once
+   */
+  const filterIngredients = useCallback(( data ) => {
+    setUserIngredients( data );
+  }, []);
 
 
   return (
@@ -58,7 +68,7 @@ const Ingredients = () => {
       <IngredientForm onAddIngredient={addIngredientHandler}/>
 
       <section>
-        <Search />
+        <Search onLoadedIngredients={filterIngredients}/>
         <IngredientList ingredients={userIngredients} onRemoveItem={removeIngredient}/>
       </section>
     </div>
